Fix #2922
The named links in list page.get_links() have no key "name" but "nameddest" instead.
When reusing an item of this list for updates we therefore must take this situation into account and fall back to "nameddest" if "name" is not present.
